+++ packages/test-menu/src/components/simple.js
@@ -1,12 +1,14 @@
 
-export default ({ show, target, h, Menu, List, didHide, getState }) =>
+export default ({ show, target, h, Menu, List, didHide, getState, height, topMenu, exposed=true }) =>
   h(Menu, {
     target,
     show,
     didHide,
     getState,
     width: 3,
-    exposed: true,
+    height,
+    topMenu,
+    exposed,
     content: h(List,
       {
         all: {
@@ -15,7 +17,7 @@ export default ({ show, target, h, Menu, List, didHide, getState }) =>
           hoverable: true,
         },
       },
-      [1,2,3,4].map(n => ({
+      [1,2,3,4,5,6,7,8].map(n => ({
         title: `Menu item ${n}`,
         key: n, // for React
       }))